Move new classes into params package
