+++ src/main/java/org/junitpioneer/jupiter/params/DisableIfDisplayName.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junitpioneer.jupiter;
+package org.junitpioneer.jupiter.params;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
